+++ src/VarNormalGamma.cpp
@@ -153,7 +153,8 @@ Rcpp::List VarNormalGammaCoefficients(Rcpp::List object) {
                             Rcpp::Named("model") = object["model"],
                             Rcpp::Named("initial") = object["initial"],
                             Rcpp::Named("priors") = object["priors"],
-                            Rcpp::Named("posterior") = write_draws(draws));
+                            Rcpp::Named("posterior") = write_draws(draws),
+                            Rcpp::Named("warnings") = reporter.warnings());
 }
 
 // [[Rcpp::export(.VarNormalGammaForecasts)]]