+++ src/VecNormalWishart.cpp
@@ -158,7 +158,8 @@ Rcpp::List VecNormalWishartCoefficients(Rcpp::List object) {
                             Rcpp::Named("model") = object["model"],
                             Rcpp::Named("initial") = object["initial"],
                             Rcpp::Named("priors") = object["priors"],
-                            Rcpp::Named("posterior") = write_draws(draws));
+                            Rcpp::Named("posterior") = write_draws(draws),
+                            Rcpp::Named("warnings") = reporter.warnings());
 }
 
 // [[Rcpp::export(.VecNormalWishartForecasts)]]