+++ src/core/models/vec_normal_gamma.cpp
@@ -30,6 +30,7 @@ using core::normalise_beta;
 using core::ssvs_sweep;
 using core::SsvsBlock;
 using core::stacked_response;
+using core::report_flat_selection_prior;
 
 VecNormalGammaDraws VecNormalGammaSampler::draw_coefficients(const VecNormalGammaInput &input,
                                                              Reporter &reporter) const
@@ -110,6 +111,8 @@ VecNormalGammaDraws VecNormalGammaSampler::draw_coefficients(const VecNormalGamm
             {
                 z_bvs = z;
                 a_bvs.emplace(input.initial.a_lambda, input.varsel_prior);
+                report_flat_selection_prior(reporter, input.spec.varsel, "a", input.varsel_prior,
+                                            input.a_prior.v_inv);
             }
         }
     }
@@ -155,6 +158,8 @@ VecNormalGammaDraws VecNormalGammaSampler::draw_coefficients(const VecNormalGamm
             if (use_bvs)
             {
                 psi_bvs.emplace(input.initial.psi_lambda, input.psi_varsel_prior);
+                report_flat_selection_prior(reporter, input.spec.varsel, "psi", input.psi_varsel_prior,
+                                            input.psi_prior.v_inv);
             }
         }
     }